Format  - some error handling
Added validation if thre is enough placeholders for variables in
formatted string. If not, a logic error is thrown. I dont check if
there is not  enough variables for  placeholders. This is not
harmfull in any way so i'll leave it as it is. This may change in  the
future.

This is the end of this task (at least for now) so I added some
note in documentation

T: C05